Fix settings.txt bug.
GraphicsSettings.INITIAL_WINDOW_WIDTH &
GraphicsSettings.INITIAL_WINDOW_HEIGHT are supposed to be int's. They
get written to ~/peril/settings.txt when it doesn't exist. Because they
were floats the client would crash because it was trying to parse the
window width & height as integers, but couldn't because they were
float's.

+++ client/src/main/java/com/forerunnergames/peril/client/settings/GraphicsSettings.java
@@ -22,8 +22,8 @@ public final class GraphicsSettings
   public static final boolean               FONT_TEXTURE_MIPMAPPING           = true;
   public static final Scaling               VIEWPORT_SCALING                  = Scaling.stretch;
 
-  public static float   INITIAL_WINDOW_WIDTH  = REFERENCE_SCREEN_WIDTH;
-  public static float   INITIAL_WINDOW_HEIGHT = REFERENCE_SCREEN_HEIGHT;
+  public static int     INITIAL_WINDOW_WIDTH  = REFERENCE_SCREEN_WIDTH;
+  public static int     INITIAL_WINDOW_HEIGHT = REFERENCE_SCREEN_HEIGHT;
   public static boolean IS_FULLSCREEN         = true;
   public static boolean IS_VSYNC_ENABLED      = true;
   public static boolean IS_WINDOW_RESIZABLE   = true;
